@h4818 2023-10-04
will be merged in master on PR https://github.com/odoo/odoo/pull/135155